Refactor custom modal to have modal reusable
